pyproject.toml
[tool.poetry]
name = "django-ocr-backend"
version = "0.1.0"
description = "The Django back end to my OCR project"
authors = ["Joseph Cardenas <josephvcardenas@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.3"
djangorestframework = "^3.12.1"
django-storages = "^1.10.1"
django-heroku = "^0.3.1"
django-cors-headers = "^3.5.0"
python-dotenv = "^0.15.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
pytesseract = "^0.3.6"
ocrmypdf = "^11.3.2"
ghostscript = "^0.6"
boto3 = "^1.16.10"
django-allauth = "^0.44.0"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
